+++ src/physic/update_entity.c
@@ -69,18 +69,18 @@ int	update_entity_against_walls(t_entity *proj, t_entity *ent, t_wall walls[64],
 	return (collision);
 }
 
-int	add_mesh(t_mesh *mesh, t_wall walls[64], int *nb_walls, int sectors_ids[16],
-	int adjacents_sectors[2], t_entity proj)
+int	add_mesh(t_mesh *mesh, t_wall walls[64], int *nb_walls, Uint32 sectors_ids[16],
+	Uint32 adjacents_sectors[2], t_entity proj)
 {
-	int		it;
+	Uint32	it;
 	t_vec4d	c;
 	t_wall	wall;
 
-	if (mesh->sector_id != -1)
+	if (mesh->sector_id != (Uint32)-1)
 	{
 		if (adjacents_sectors[0] == 16)
 			return (-1);
-		it = -1;
+		it = (Uint32)-1;
 		while (++it < adjacents_sectors[0])
 			if (sectors_ids[it] == mesh->sector_id)
 				break ;
@@ -90,7 +90,7 @@ int	add_mesh(t_mesh *mesh, t_wall walls[64], int *nb_walls, int sectors_ids[16],
 			sectors_ids[it] = mesh->sector_id;
 		}
 	}
-	it = -1;
+	it = (Uint32)-1;
 	while (++it < mesh->nb_walls)
 	{
 		c = (t_vec4d){.c3 = {.vec3d = mesh->walls[it].center, .w = 1}};
@@ -112,19 +112,19 @@ int	add_mesh(t_mesh *mesh, t_wall walls[64], int *nb_walls, int sectors_ids[16],
 int	prepare_walls(t_wall walls[64], t_entity proj, t_sector *sector,
 	t_world *world)
 {
-	int	nb_walls;
-	int	sector_ids[16];
-	int	adjacent_sectors[2];
-	int	it;
+	int		nb_walls;
+	Uint32	sector_ids[16];
+	Uint32	adjacent_sectors[2];
+	Uint32	it;
 
 	nb_walls = 0;
-	sector_ids[0] = (int)((sector - world->sectors) / sizeof(t_sector));
+	sector_ids[0] = (Uint32)((sector - world->sectors) / sizeof(t_sector));
 	adjacent_sectors[0] = 1;
 	adjacent_sectors[1] = 0;
 	while (adjacent_sectors[1] < adjacent_sectors[0])
 	{
 		sector = &world->sectors[sector_ids[adjacent_sectors[1]++]];
-		it = -1;
+		it = (Uint32)-1;
 		while (++it < sector->meshnum)
 			add_mesh(&sector->mesh[it], walls, &nb_walls, sector_ids,
 				adjacent_sectors, proj);
